cheri: Fixes required for SM768
The introduction of the SM768 driver triggered several
fixes that the generally useful.

+++ drivers/video/fbdev/core/cfbimgblt.c
@@ -126,7 +126,7 @@ static inline void color_imageblit(const struct fb_image *image,
 		dst1 += p->fix.line_length;
 		if (pitch_index) {
 			dst2 += p->fix.line_length;
-			dst1 = (u8 __iomem *)((long __force)dst2 & ~(sizeof(u32) - 1));
+			dst1 = (u8 __iomem *)((uintptr_t __force)dst2 & ~(sizeof(u32) - 1));
 
 			start_index += pitch_index;
 			start_index &= 32 - 1;
@@ -196,7 +196,7 @@ static inline void slow_imageblit(const struct fb_image *image, struct fb_info *
 		src += spitch;
 		if (pitch_index) {
 			dst2 += pitch;
-			dst1 = (u8 __iomem *)((long __force)dst2 & ~(sizeof(u32) - 1));
+			dst1 = (u8 __iomem *)((uintptr_t __force)dst2 & ~(sizeof(u32) - 1));
 			start_index += pitch_index;
 			start_index &= 32 - 1;
 		}
